[AI] Move i18n usage outside of loot-core
loot-core should be platform-agnostic and free of i18n concerns. This
moves all user-facing translated string helpers (rule/schedule labels,
error formatters) into desktop-client's #util/{rule,schedule,error}
modules, while keeping the underlying logic in loot-core.

- Headless @actual-app/api error formatters in loot-core now return
  plain English strings (the API has no i18n runtime).
- The persisted "Unknown" institution name and the platform storage
  alert use plain strings.
- Removed the unused i18next dependency from loot-core.
