+++ packages/desktop-client/src/app/appSlice.ts
@@ -1,5 +1,4 @@
 import { send } from '@actual-app/core/platform/client/connection';
-import { getUploadError } from '@actual-app/core/shared/errors';
 import type { AtLeastOne } from '@actual-app/core/types/util';
 import { createAction, createSlice } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
@@ -7,6 +6,7 @@ import type { PayloadAction } from '@reduxjs/toolkit';
 import { pushModal } from '#modals/modalsSlice';
 import { loadPrefs } from '#prefs/prefsSlice';
 import { createAppAsyncThunk } from '#redux';
+import { getUploadError } from '#util/error';
 import { getIsOutdated, getLatestVersion } from '#util/versions';
 
 const sliceName = 'app';
