+++ packages/desktop-client/src/components/admin/UserAccess/UserAccessRow.tsx
@@ -5,7 +5,6 @@ import { useTranslation } from 'react-i18next';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 import { send } from '@actual-app/core/platform/client/connection';
-import { getUserAccessErrors } from '@actual-app/core/shared/errors';
 import type { UserAvailable } from '@actual-app/core/types/models';
 
 import { Checkbox } from '#components/forms';
@@ -14,6 +13,7 @@ import { useMetadataPref } from '#hooks/useMetadataPref';
 import { addNotification } from '#notifications/notificationsSlice';
 import { useDispatch } from '#redux';
 import { signOut } from '#users/usersSlice';
+import { getUserAccessErrors } from '#util/error';
 
 type UserAccessProps = {
   access: UserAvailable;
