+++ packages/desktop-client/src/components/modals/EnableBankingInitialiseModal.tsx
@@ -10,7 +10,6 @@ import { Text } from '@actual-app/components/text';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 import { send } from '@actual-app/core/platform/client/connection';
-import { getSecretsError } from '@actual-app/core/shared/errors';
 
 import { Error as ErrorAlert } from '#components/alerts';
 import { Link } from '#components/common/Link';
@@ -22,6 +21,7 @@ import {
 } from '#components/common/Modal';
 import { FormField, FormLabel } from '#components/forms';
 import type { Modal as ModalType } from '#modals/modalsSlice';
+import { getSecretsError } from '#util/error';
 
 type EnableBankingInitialiseProps = Extract<
   ModalType,
