+++ packages/desktop-client/src/components/modals/SimpleFinInitialiseModal.tsx
@@ -7,7 +7,6 @@ import { Input } from '@actual-app/components/input';
 import { Text } from '@actual-app/components/text';
 import { View } from '@actual-app/components/view';
 import { send } from '@actual-app/core/platform/client/connection';
-import { getSecretsError } from '@actual-app/core/shared/errors';
 
 import { Error } from '#components/alerts';
 import { Link } from '#components/common/Link';
@@ -19,6 +18,7 @@ import {
 } from '#components/common/Modal';
 import { FormField, FormLabel } from '#components/forms';
 import type { Modal as ModalType } from '#modals/modalsSlice';
+import { getSecretsError } from '#util/error';
 
 type SimpleFinInitialiseModalProps = Extract<
   ModalType,
