+++ packages/desktop-client/src/components/reports/reports/Sankey.tsx
@@ -20,7 +20,6 @@ import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
 import { send } from '@actual-app/core/platform/client/connection';
 import * as monthUtils from '@actual-app/core/shared/months';
-import { mapField } from '@actual-app/core/shared/rules';
 import type {
   RuleConditionEntity,
   SankeyWidget,
@@ -58,6 +57,7 @@ import { useRuleConditionFilters } from '#hooks/useRuleConditionFilters';
 import { addNotification } from '#notifications/notificationsSlice';
 import { useDispatch } from '#redux';
 import { useUpdateDashboardWidgetMutation } from '#reports/mutations';
+import { mapField } from '#util/rule';
 
 export function Sankey() {
   const params = useParams();
