+++ packages/desktop-client/src/components/rules/ConditionExpression.tsx
@@ -4,7 +4,8 @@ import type { CSSProperties } from 'react';
 import { Text } from '@actual-app/components/text';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
-import { friendlyOp, mapField } from '@actual-app/core/shared/rules';
+
+import { friendlyOp, mapField } from '#util/rule';
 
 import { Value } from './Value';
