+++ packages/desktop-client/src/components/rules/RuleEditor.tsx
@@ -27,13 +27,10 @@ import * as monthUtils from '@actual-app/core/shared/months';
 import { q } from '@actual-app/core/shared/query';
 import {
   FIELD_TYPES,
-  friendlyOp,
-  getAllocationMethods,
   getFieldError,
   getValidOps,
   isValidOp,
   makeValue,
-  mapField,
   parse,
   unparse,
 } from '@actual-app/core/shared/rules';
@@ -62,6 +59,7 @@ import { addNotification } from '#notifications/notificationsSlice';
 import { aqlQuery } from '#queries/aqlQuery';
 import { useDispatch } from '#redux';
 import { disableUndo, enableUndo } from '#undo';
+import { friendlyOp, getAllocationMethods, mapField } from '#util/rule';
 
 import { FormulaActionEditor } from './FormulaActionEditor';
