+++ packages/desktop-client/src/components/rules/RuleRow.tsx
@@ -11,13 +11,16 @@ import { styles } from '@actual-app/components/styles';
 import { Text } from '@actual-app/components/text';
 import { theme } from '@actual-app/components/theme';
 import { View } from '@actual-app/components/view';
-import { friendlyOp, translateRuleStage } from '@actual-app/core/shared/rules';
 import type { RuleEntity } from '@actual-app/core/types/models';
 
 import { Cell, Field, Row, SelectCell } from '#components/table';
 import { useContextMenu } from '#hooks/useContextMenu';
 import { useSelectedDispatch } from '#hooks/useSelected';
-import { groupActionsBySplitIndex } from '#util/ruleUtils';
+import {
+  friendlyOp,
+  groupActionsBySplitIndex,
+  translateRuleStage,
+} from '#util/rule';
 
 import { ActionExpression } from './ActionExpression';
 import { ConditionExpression } from './ConditionExpression';
