+++ packages/loot-core/package.json
@@ -162,7 +162,6 @@
   "scripts": {
     "build:node": "cross-env NODE_ENV=production vite build --config ./vite.desktop.config.mts",
     "watch:node": "cross-env NODE_ENV=development vite build --config ./vite.desktop.config.mts --watch",
-    "generate:i18n": "i18next",
     "test": "npm-run-all -cp 'test:*'",
     "test:node": "ENV=node vitest --run",
     "test:web": "ENV=web vitest --run -c vitest.web.config.ts",
@@ -207,7 +206,6 @@
     "crypto-browserify": "^3.12.1",
     "fake-indexeddb": "^6.2.5",
     "fast-check": "^4.6.0",
-    "i18next": "^25.10.10",
     "jest-diff": "^30.3.0",
     "jsverify": "^0.8.4",
     "mockdate": "^3.0.5",
