+++ packages/loot-core/src/server/accounts/app.ts
@@ -1,4 +1,3 @@
-import { t } from 'i18next';
 import { v4 as uuidv4 } from 'uuid';
 
 import { captureException } from '#platform/exceptions';
@@ -237,7 +236,7 @@ async function linkSimpleFinAccount({
   let id;
 
   const institution = {
-    name: externalAccount.institution ?? t('Unknown'),
+    name: externalAccount.institution ?? 'Unknown',
   };
 
   const bank = await link.findOrCreateBank(
@@ -311,7 +310,7 @@ async function linkPluggyAiAccount({
   let id;
 
   const institution = {
-    name: externalAccount.institution ?? t('Unknown'),
+    name: externalAccount.institution ?? 'Unknown',
   };
 
   const bank = await link.findOrCreateBank(
@@ -385,7 +384,7 @@ async function linkEnableBankingAccount({
   let id: string | undefined;
 
   const institution = {
-    name: externalAccount.institution ?? t('Unknown'),
+    name: externalAccount.institution ?? 'Unknown',
   };
 
   // Enable Banking uses a session-per-account model, so we use the
