+++ upcoming-release-notes/8027.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [MatissJanis]
+---
+
+Move i18n usage outside of loot-core.
